+++ R/generate_pseudo_population.R
@@ -18,6 +18,8 @@
 #'   - "parallel": Uses parallel flag whenever possible. (Currently is activated
 #'    on the SuperLearner Module.)
 #' @param pred_model a prediction model
+#' @param gps_model Model type which is used for estimating GPS value, including
+#' parametric (default) and non-parametric.
 #' @param use_cov_transform If TRUE, the function uses transformer to meet the
 #'  covariate balance.
 #' @param transformers A list of transformers. Is transformer should be a
@@ -25,6 +27,9 @@
 #' Available transformers:
 #'   - pow2: to the power of 2
 #'   - pow3: to the power of 3#'
+#' @param bin_seq Sequence of w (treatment) to generate pseudo population. If
+#' NULL is passed the default value will be used, which is
+#' `seq(min(w)+delta_n/2,max(w), by=delta_n)`.
 #' @param save_output If TRUE, output results will be stored at the save.path.
 #'  Default is FALSE.
 #' @param save_path location for storing the final results, format of the saved
@@ -98,8 +103,10 @@ gen_pseudo_pop <- function(Y,
                            ci_appr,
                            running_appr,
                            pred_model,
+                           gps_model = "parametric",
                            use_cov_transform = FALSE,
                            transformers = list("pow2","pow3"),
+                           bin_seq = NULL,
                            save_output = FALSE,
                            save_path = NULL,
                            params = list(),
@@ -114,13 +121,12 @@ gen_pseudo_pop <- function(Y,
   # timing the function
   st_time_gpp <- proc.time()
 
-
   # function call
   fcall <- match.call()
 
   # Check arguments ----------------------------------------
   check_args(pred_model,ci_appr,running_appr, use_cov_transform, transformers,
-             ...)
+             gps_model, ...)
 
   # Generate output set ------------------------------------
   counter <- 0
@@ -138,15 +144,14 @@ gen_pseudo_pop <- function(Y,
   # do not use gps values.
   # TODO: find a better place to the following code.
   tmp_data <- cbind(Y,w,w,c)
-
-  q1 <- quantile(tmp_data$w,0.01)
-  q2 <- quantile(tmp_data$w,0.99)
-
-  tmp_data <- subset(tmp_data[complete.cases(tmp_data) ,],  w < q2)
-  tmp_data <- subset(tmp_data[complete.cases(tmp_data) ,],  w < q2  & w > q1)
-
+  q1 <- stats::quantile(tmp_data$w,0.01)
+  q2 <- stats::quantile(tmp_data$w,0.99)
+  tmp_data <- subset(tmp_data[stats::complete.cases(tmp_data) ,],  w < q2  & w > q1)
   original_corr_obj <- check_covar_balance(tmp_data, ci_appr, nthread, ...)
   tmp_data <- NULL
+
+  logger::log_debug("1% qauntile for trim: {q1}")
+  logger::log_debug("99% qauntile for trim: {q2}")
 
   # loop until the generated pseudo population is acceptable or reach maximum
   # allowed iteration.
@@ -170,7 +175,8 @@ gen_pseudo_pop <- function(Y,
 
     ## Estimate GPS -----------------------------
     logger::log_debug("Started to estimate gps ... ")
-    estimate_gps_out <- estimate_gps(Y, w, c_extended[unlist(covariate_cols)], pred_model, running_appr,
+    estimate_gps_out <- estimate_gps(Y, w, c_extended[unlist(covariate_cols)],
+                                     pred_model, running_appr, gps_model,
                                      params = params, nthread = nthread,
                                      internal_use = internal_use, ...)
     logger::log_debug("Finished estimating gps.")
@@ -190,10 +196,10 @@ gen_pseudo_pop <- function(Y,
 
     ## Compile data ---------
     logger::log_debug("Started compiling pseudo population ... ")
-    pseudo_pop <- compile_pseudo_pop(dataset=estimate_gps_out,
-                                     ci_appr=ci_appr, nthread = nthread, ...)
-    pseudo_pop <- subset(pseudo_pop[complete.cases(pseudo_pop) ,],  w < q2)
-    pseudo_pop <- subset(pseudo_pop[complete.cases(pseudo_pop) ,],  w < q2  & w > q1)
+    pseudo_pop <- compile_pseudo_pop(dataset=estimate_gps_out, ci_appr=ci_appr,
+                                     gps_model,bin_seq, nthread = nthread, ...)
+    pseudo_pop <- subset(pseudo_pop[stats::complete.cases(pseudo_pop) ,],
+                         w < q2  & w > q1)
     logger::log_debug("Finished compiling pseudo population.")
 
     if (ci_appr == 'adjust'){
@@ -214,7 +220,6 @@ gen_pseudo_pop <- function(Y,
       best_pseudo_pop <- pseudo_pop
       best_adjusted_corr_obj <- adjusted_corr_obj
     }
-
 
     if (adjusted_corr_obj$pass){
       message(paste('Covariate balance condition has been met (iteration: ',
@@ -253,7 +258,6 @@ gen_pseudo_pop <- function(Y,
             value_found = TRUE
             break
           }
-
         }
         if (value_found){break}
       }